Refactor harn-vm VM module layout
