+++ crates/harn-vm/src/vm/execution.rs
@@ -1,0 +1,306 @@
+use std::rc::Rc;
+use std::time::Instant;
+
+use crate::chunk::Chunk;
+use crate::value::{ModuleFunctionRegistry, VmError, VmValue};
+
+use super::{CallFrame, Vm};
+
+impl Vm {
+    /// Execute a compiled chunk.
+    pub async fn execute(&mut self, chunk: &Chunk) -> Result<VmValue, VmError> {
+        let span_id = crate::tracing::span_start(crate::tracing::SpanKind::Pipeline, "main".into());
+        let result = self.run_chunk(chunk).await;
+        crate::tracing::span_end(span_id);
+        result
+    }
+
+    /// Convert a VmError into either a handled exception (returning Ok) or a propagated error.
+    pub(crate) fn handle_error(&mut self, error: VmError) -> Result<Option<VmValue>, VmError> {
+        let thrown_value = match &error {
+            VmError::Thrown(v) => v.clone(),
+            other => VmValue::String(Rc::from(other.to_string())),
+        };
+
+        if let Some(handler) = self.exception_handlers.pop() {
+            if !handler.error_type.is_empty() {
+                // Typed catch: only match when the thrown enum's type equals the declared type.
+                let matches = match &thrown_value {
+                    VmValue::EnumVariant { enum_name, .. } => *enum_name == handler.error_type,
+                    _ => false,
+                };
+                if !matches {
+                    return self.handle_error(error);
+                }
+            }
+
+            while self.frames.len() > handler.frame_depth {
+                if let Some(frame) = self.frames.pop() {
+                    if let Some(ref dir) = frame.saved_source_dir {
+                        crate::stdlib::set_thread_source_dir(dir);
+                    }
+                    self.iterators.truncate(frame.saved_iterator_depth);
+                    self.env = frame.saved_env;
+                }
+            }
+
+            // Drop deadlines that belonged to unwound frames.
+            while self
+                .deadlines
+                .last()
+                .is_some_and(|d| d.1 > handler.frame_depth)
+            {
+                self.deadlines.pop();
+            }
+
+            self.env.truncate_scopes(handler.env_scope_depth);
+
+            self.stack.truncate(handler.stack_depth);
+            self.stack.push(thrown_value);
+
+            if let Some(frame) = self.frames.last_mut() {
+                frame.ip = handler.catch_ip;
+            }
+
+            Ok(None)
+        } else {
+            Err(error)
+        }
+    }
+
+    pub(crate) async fn run_chunk(&mut self, chunk: &Chunk) -> Result<VmValue, VmError> {
+        self.run_chunk_entry(chunk, 0, None, None, None).await
+    }
+
+    pub(crate) async fn run_chunk_entry(
+        &mut self,
+        chunk: &Chunk,
+        argc: usize,
+        saved_source_dir: Option<std::path::PathBuf>,
+        module_functions: Option<ModuleFunctionRegistry>,
+        module_state: Option<crate::value::ModuleState>,
+    ) -> Result<VmValue, VmError> {
+        let initial_env = self.env.clone();
+        self.frames.push(CallFrame {
+            chunk: chunk.clone(),
+            ip: 0,
+            stack_base: self.stack.len(),
+            saved_env: self.env.clone(),
+            initial_env: Some(initial_env),
+            saved_iterator_depth: self.iterators.len(),
+            fn_name: String::new(),
+            argc,
+            saved_source_dir,
+            module_functions,
+            module_state,
+        });
+
+        loop {
+            if let Some(&(deadline, _)) = self.deadlines.last() {
+                if Instant::now() > deadline {
+                    self.deadlines.pop();
+                    let err = VmError::Thrown(VmValue::String(Rc::from("Deadline exceeded")));
+                    match self.handle_error(err) {
+                        Ok(None) => continue,
+                        Ok(Some(val)) => return Ok(val),
+                        Err(e) => return Err(e),
+                    }
+                }
+            }
+
+            let frame = match self.frames.last_mut() {
+                Some(f) => f,
+                None => return Ok(self.stack.pop().unwrap_or(VmValue::Nil)),
+            };
+
+            if frame.ip >= frame.chunk.code.len() {
+                let val = self.stack.pop().unwrap_or(VmValue::Nil);
+                let popped_frame = self.frames.pop().unwrap();
+                if let Some(ref dir) = popped_frame.saved_source_dir {
+                    crate::stdlib::set_thread_source_dir(dir);
+                }
+
+                if self.frames.is_empty() {
+                    return Ok(val);
+                } else {
+                    self.iterators.truncate(popped_frame.saved_iterator_depth);
+                    self.env = popped_frame.saved_env;
+                    self.stack.truncate(popped_frame.stack_base);
+                    self.stack.push(val);
+                    continue;
+                }
+            }
+
+            let op = frame.chunk.code[frame.ip];
+            frame.ip += 1;
+
+            match self.execute_op(op).await {
+                Ok(Some(val)) => return Ok(val),
+                Ok(None) => continue,
+                Err(VmError::Return(val)) => {
+                    if let Some(popped_frame) = self.frames.pop() {
+                        if let Some(ref dir) = popped_frame.saved_source_dir {
+                            crate::stdlib::set_thread_source_dir(dir);
+                        }
+                        let current_depth = self.frames.len();
+                        self.exception_handlers
+                            .retain(|h| h.frame_depth <= current_depth);
+
+                        if self.frames.is_empty() {
+                            return Ok(val);
+                        }
+                        self.iterators.truncate(popped_frame.saved_iterator_depth);
+                        self.env = popped_frame.saved_env;
+                        self.stack.truncate(popped_frame.stack_base);
+                        self.stack.push(val);
+                    } else {
+                        return Ok(val);
+                    }
+                }
+                Err(e) => {
+                    // Capture stack trace before error handling unwinds frames.
+                    if self.error_stack_trace.is_empty() {
+                        self.error_stack_trace = self.capture_stack_trace();
+                    }
+                    match self.handle_error(e) {
+                        Ok(None) => {
+                            self.error_stack_trace.clear();
+                            continue;
+                        }
+                        Ok(Some(val)) => return Ok(val),
+                        Err(e) => return Err(self.enrich_error_with_line(e)),
+                    }
+                }
+            }
+        }
+    }
+
+    pub(crate) async fn execute_one_cycle(&mut self) -> Result<Option<(VmValue, bool)>, VmError> {
+        if let Some(&(deadline, _)) = self.deadlines.last() {
+            if Instant::now() > deadline {
+                self.deadlines.pop();
+                let err = VmError::Thrown(VmValue::String(Rc::from("Deadline exceeded")));
+                match self.handle_error(err) {
+                    Ok(None) => return Ok(None),
+                    Ok(Some(val)) => return Ok(Some((val, false))),
+                    Err(e) => return Err(e),
+                }
+            }
+        }
+
+        let frame = match self.frames.last_mut() {
+            Some(f) => f,
+            None => {
+                let val = self.stack.pop().unwrap_or(VmValue::Nil);
+                return Ok(Some((val, false)));
+            }
+        };
+
+        if frame.ip >= frame.chunk.code.len() {
+            let val = self.stack.pop().unwrap_or(VmValue::Nil);
+            let popped_frame = self.frames.pop().unwrap();
+            if self.frames.is_empty() {
+                return Ok(Some((val, false)));
+            } else {
+                self.iterators.truncate(popped_frame.saved_iterator_depth);
+                self.env = popped_frame.saved_env;
+                self.stack.truncate(popped_frame.stack_base);
+                self.stack.push(val);
+                return Ok(None);
+            }
+        }
+
+        let op = frame.chunk.code[frame.ip];
+        frame.ip += 1;
+
+        match self.execute_op(op).await {
+            Ok(Some(val)) => Ok(Some((val, false))),
+            Ok(None) => Ok(None),
+            Err(VmError::Return(val)) => {
+                if let Some(popped_frame) = self.frames.pop() {
+                    if let Some(ref dir) = popped_frame.saved_source_dir {
+                        crate::stdlib::set_thread_source_dir(dir);
+                    }
+                    let current_depth = self.frames.len();
+                    self.exception_handlers
+                        .retain(|h| h.frame_depth <= current_depth);
+                    if self.frames.is_empty() {
+                        return Ok(Some((val, false)));
+                    }
+                    self.iterators.truncate(popped_frame.saved_iterator_depth);
+                    self.env = popped_frame.saved_env;
+                    self.stack.truncate(popped_frame.stack_base);
+                    self.stack.push(val);
+                    Ok(None)
+                } else {
+                    Ok(Some((val, false)))
+                }
+            }
+            Err(e) => {
+                if self.error_stack_trace.is_empty() {
+                    self.error_stack_trace = self.capture_stack_trace();
+                }
+                match self.handle_error(e) {
+                    Ok(None) => {
+                        self.error_stack_trace.clear();
+                        Ok(None)
+                    }
+                    Ok(Some(val)) => Ok(Some((val, false))),
+                    Err(e) => Err(self.enrich_error_with_line(e)),
+                }
+            }
+        }
+    }
+
+    /// Capture the current call stack as (fn_name, line, col, source_file) tuples.
+    pub(crate) fn capture_stack_trace(&self) -> Vec<(String, usize, usize, Option<String>)> {
+        self.frames
+            .iter()
+            .map(|f| {
+                let idx = if f.ip > 0 { f.ip - 1 } else { 0 };
+                let line = f.chunk.lines.get(idx).copied().unwrap_or(0) as usize;
+                let col = f.chunk.columns.get(idx).copied().unwrap_or(0) as usize;
+                (f.fn_name.clone(), line, col, f.chunk.source_file.clone())
+            })
+            .collect()
+    }
+
+    /// Enrich a VmError with source line information from the captured stack
+    /// trace. Appends ` (line N)` to error variants whose messages don't
+    /// already carry location context.
+    pub(crate) fn enrich_error_with_line(&self, error: VmError) -> VmError {
+        // Determine the line from the captured stack trace (innermost frame).
+        let line = self
+            .error_stack_trace
+            .last()
+            .map(|(_, l, _, _)| *l)
+            .unwrap_or_else(|| self.current_line());
+        if line == 0 {
+            return error;
+        }
+        let suffix = format!(" (line {line})");
+        match error {
+            VmError::Runtime(msg) => VmError::Runtime(format!("{msg}{suffix}")),
+            VmError::TypeError(msg) => VmError::TypeError(format!("{msg}{suffix}")),
+            VmError::DivisionByZero => VmError::Runtime(format!("Division by zero{suffix}")),
+            VmError::UndefinedVariable(name) => {
+                VmError::Runtime(format!("Undefined variable: {name}{suffix}"))
+            }
+            VmError::UndefinedBuiltin(name) => {
+                VmError::Runtime(format!("Undefined builtin: {name}{suffix}"))
+            }
+            VmError::ImmutableAssignment(name) => VmError::Runtime(format!(
+                "Cannot assign to immutable binding: {name}{suffix}"
+            )),
+            VmError::StackOverflow => {
+                VmError::Runtime(format!("Stack overflow: too many nested calls{suffix}"))
+            }
+            // Leave these untouched:
+            // - Thrown: user-thrown errors should not be silently modified
+            // - CategorizedError: structured errors for agent orchestration
+            // - Return: control flow, not a real error
+            // - StackUnderflow / InvalidInstruction: internal VM bugs
+            other => other,
+        }
+    }
+}
