+++ crates/harn-vm/src/vm/mod.rs
@@ -1,0 +1,26 @@
+mod async_builtin;
+mod debug;
+mod dispatch;
+mod execution;
+mod format;
+pub mod iter;
+mod methods;
+mod modules;
+mod ops;
+mod scope;
+mod state;
+
+#[cfg(test)]
+mod tests_debug;
+#[cfg(test)]
+mod tests_runtime;
+
+#[allow(deprecated)]
+pub use async_builtin::{
+    clone_async_builtin_child_vm, install_async_builtin_child_vm, restore_async_builtin_child_vm,
+    take_async_builtin_child_vm, AsyncBuiltinChildVmGuard,
+};
+pub use debug::{DebugAction, DebugState};
+pub use state::Vm;
+
+pub(crate) use state::{CallFrame, ExceptionHandler, IterState, ScopeSpan};
